StackOverflow/requirement.txt
Django
django-rest-framework
djangorestframework-simplejwt
psycopg2
celery